refactor library creation
