+++ CMakeLists.txt
@@ -10,11 +10,10 @@ if(${CMAKE_VERSION} VERSION_LESS 3.12)
     cmake_policy(VERSION ${CMAKE_MAJOR_VERSION}.${CMAKE_MINOR_VERSION})
 endif()
 
+add_subdirectory(src)
+
 add_executable(
-    station 
-    src/main.cpp
-    src/cpp/part.cpp
-    src/cpp/ships.cpp
-    src/cpp/repairBay.cpp
-    src/cpp/station.cpp
+    station-zebra 
+    main.cpp
 )
+target_link_libraries(station-zebra PRIVATE station)
